requirements.txt
aiogram==2.25.1
PyYAML==6.0.1
loguru
requests
windows-curses